Each class has own file in auxiliary_application
